Changement des types des entités

+++ src/main/PacmanGame/AI/OrangeGhost.java
@@ -9,7 +9,7 @@ public class OrangeGhost extends Ghost {
     private double initialX;
     private double initialY;
 
-    public OrangeGhost(int initialX, int initialY, CoreEngine coreEngine) {
+    public OrangeGhost(double initialX, double initialY, CoreEngine coreEngine) {
         super(initialX, initialY, coreEngine);
     }
